+++ app/src/src_nofdroid/org/breezyweather/sources/eccc/EcccResultConverter.kt
@@ -175,7 +175,8 @@ private fun getDailyForecast(
                             precipitationProbability = PrecipitationProbability(
                                 total = nighttime.precip?.toDoubleOrNull()
                             )
-                        )
+                        ),
+                        sunshineDuration = daytime?.sun?.value?.toDoubleOrNull()
                     )
                 )
             }