+++ app/src/src_nofdroid/org/breezyweather/sources/eccc/json/EcccSun.kt
@@ -17,19 +17,10 @@
 package org.breezyweather.sources.eccc.json
 
 import kotlinx.serialization.Serializable
+import org.breezyweather.sources.eccc.EcccSunSerializer
 
-@Serializable
-data class EcccDaily(
-    val date: String?,
-    val summary: String?,
-    val periodLabel: String?,
-
-    val iconCode: String?,
-    val text: String?,
-
-    val temperature: EcccDailyTemperature?,
-
-    val precip: String?,
-
-    val sun: EcccSun?
+@Serializable(with = EcccSunSerializer::class)
+data class EcccSun(
+    val value: String?,
+    //val unit: String? // Always "hours"
 )